Prevent parser crashes on unsupported Lottie elements and sparse JSON
Safely handle unrecognized layer and shape discriminators and missing schema properties during Lottie JSON decoding. Unrecognized shapes and layers deserialize into safe UnknownElement and UnknownLayer AST nodes rather than failing with MissingFieldException or throwing deserializer errors.

### Motivation
Real-world Lottie exports often contain elements outside the baseline specification (such as audio, text, or unsupported modifiers) as well as sparse property sets where optional or defaulted values like frame bounds or nested shape lists are omitted. Falling back to Group or failing prematurely crashes the decoder and blocks rendering of valid sibling elements. Preserving unknown layers and elements in the AST allows transform chains and layer parenting to function correctly while renderer passes safely skip unknown nodes.

+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/format/layer/NullLayer.kt
@@ -28,7 +28,7 @@ internal data class NullLayer(
   @SerialName("ty") override val type: LayerType = LayerType.Null,
   @SerialName("ind") override val index: Int? = null,
   @SerialName("parent") override val parent: Int? = null,
-  @SerialName("ip") override val startFrame: Int? = null,
-  @SerialName("op") override val endFrame: Int? = null,
+  @SerialName("ip") override val startFrame: Float? = null,
+  @SerialName("op") override val endFrame: Float? = null,
   @SerialName("ks") override val transform: Transform? = null,
 ) : Layer()
